Add toggle for captive portal check
In order to detect working data, the portal check connects to
google servers.
But in some countries, for example china, google servers are
blocked and the check was never successful.

This change adds a toggle for users, who are in a restricted
network which blocks "clients3.google.com" and allows them
to use their network as usual

Change-Id: I2cb75a3234c1df894c238ed6b5e37495434a417f
Signed-off-by: Alexander Martinz <eviscerationls@gmail.com>

Conflicts:
	res/values/slim_strings.xml